% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{capital_remaining}
\alias{capital_remaining}
\title{capital_remaining}
\usage{
capital_remaining(rate, capital, term_months, month_num, month_overpay = 0)
}
\arguments{
\item{rate}{double, the mortgage current rate (absolute value e.g. 1.5\% means rate = 0.015)}

\item{capital}{double, the mortgage remaining capital}

\item{term_months}{integer, number of months remaining to pay the mortgage back}

\item{month_num}{integer, duration of payment period, in months, payments equal those required to pay off mortgage without overpayments}

\item{month_overpay}{integer, amount overpaid each month, defaults to 0}
}
\value{
double capital remaining after month_num
}
\description{
capital remaining after paying for month_num months
}
