% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{interest_paid_splitloan_diffterm}
\alias{interest_paid_splitloan_diffterm}
\title{interest_paid_splitloan_diffterm}
\usage{
interest_paid_splitloan_diffterm(
  rate,
  capital1,
  capital2,
  term_months1,
  term_months2,
  month_num,
  month_overpay = rep(0, month_num),
  perc_overpay1
)
}
\arguments{
\item{rate}{double, the mortgage current rate (absolute value e.g. 1.5\% means rate = 0.015)}

\item{capital1}{double, the mortgage remaining capital1}

\item{capital2}{double, the mortgage remaining capital2}

\item{term_months1}{integer, number of months remaining to pay the 1st mortgage back}

\item{term_months2}{integer, number of months remaining to pay the 2nd mortgage back}

\item{month_num}{integer, duration of payment period, in months, payments equal those required to pay off mortgage without overpayments}

\item{month_overpay}{integer, amount overpaid each month, defaults to 0}

\item{perc_overpay1}{percentage of overpay dedicated to first loan}
}
\value{
named list reflecting state of both mortgage, with members capital1, capital2, act_month1, act_month2, int_paid1, int_paid2, int_paid
}
\description{
interest paid on two loans of same rate but different terms
}
