% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{monthly_payment}
\alias{monthly_payment}
\title{monthly_payment}
\usage{
monthly_payment(rate, capital, term_months, rest = 12)
}
\arguments{
\item{rate}{double, the mortgage current rate (absolute value e.g. 1.5\% means rate = 0.015)}

\item{capital}{double, the mortgage remaining capital}

\item{term_months}{integer, number of months remaining to pay the mortgage back integer, number of months remaining to pay the mortgage back}

\item{rest}{integer, defaults to 12 for monthly payments}
}
\value{
double monthly payment
}
\description{
monthly payment for a classic mortgage
}
