% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mortgage_savings_comparison}
\alias{mortgage_savings_comparison}
\title{mortgage_savings_comparison}
\usage{
mortgage_savings_comparison(
  rate,
  capital,
  term_months,
  savings,
  flat_value,
  down_payment,
  month_num,
  month_overpay_vect,
  reduce_term = TRUE,
  case_name = "default"
)
}
\arguments{
\item{rate}{vector double, the mortgage rates (absolute value e.g. 1.5\% means rate = 0.015)}

\item{capital}{vector double, the mortgage remaining capital}

\item{term_months}{vector integer, number of months remaining to pay the mortgage back without overpayments}

\item{savings}{integer, amount in savings prior to duration}

\item{flat_value}{integer, value of the flat}

\item{down_payment}{vector integer, amount to take from savings and overpay ahead of loan}

\item{month_num}{integer, duration of payment period considered, in months}

\item{month_overpay_vect}{list vector double, amount overpaid each month, defaults to rep(0,month_num)}

\item{reduce_term}{vector boolean, defaults to TRUE do overpayments reduce terms}

\item{case_name}{vector string, defaults to "default" where the case_names will be "case_i"}
}
\value{
dataframe with each observation being case and
variable including case_name, rate, capital_start, reduce_term, monthly_payment_tot, interest_paid_tot and capital_end
}
\description{
gives summary of mortgage costs considering parameters, as a dataframe
}
