% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mortgage_savings_fees_comparison_grid}
\alias{mortgage_savings_fees_comparison_grid}
\title{mortgage_savings_fees_comparison_grid}
\usage{
mortgage_savings_fees_comparison_grid(
  rate,
  fees,
  add_fees_to_loan,
  capital,
  fixed_term_years = NA,
  term_months,
  savings,
  flat_value,
  down_payment,
  month_num,
  month_overpay_vect,
  reduce_term = TRUE,
  case_name_fun = NULL,
  normalize_cashflow = TRUE,
  max_monthly_outgoing = Inf,
  products = NULL
)
}
\arguments{
\item{rate}{vector double, the mortgage rates (absolute value e.g. 1.5\% means rate = 0.015)}

\item{fees}{vector double, the mortgage fees}

\item{add_fees_to_loan}{vector boolean, whether to add the fees to the loan (TRUE) or pay it off as a single down payment (FALSE)}

\item{capital}{vector double, the mortgage remaining capital}

\item{fixed_term_years}{vector integer, number of years fixed term assured}

\item{term_months}{vector integer, number of months remaining to pay the mortgage back without overpayments}

\item{savings}{integer, amount in savings prior to duration}

\item{flat_value}{integer, value of the flat}

\item{down_payment}{vector integer, amount to take from savings and overpay ahead of loan}

\item{month_num}{integer, duration of payment period considered, in months}

\item{month_overpay_vect}{list vector double, amount overpaid each month, defaults to rep(0,month_num)}

\item{reduce_term}{vector boolean, defaults to TRUE do overpayments reduce terms}

\item{case_name_fun}{optional function to derive the case name, defaults to a concatenation of relevant parameters}

\item{normalize_cashflow}{optional boolean whether to normalize the cashflow}

\item{max_monthly_outgoing}{optional double maximum amount payable per month, cases greater than the value will be removed with a warning
max_monthly_outgoing}

\item{products}{optional data.frame bundling linked product params with the required columns being (product_name, rate, fees, fixed_term_years). if provided, rate, fees from the parameters will be ignored.}
}
\value{
dataframe with each observation being case and
variable including case_name, rate, capital_start, reduce_term, monthly_payment_tot, interest_paid_tot and capital_end
}
\description{
gives summary of mortgage costs considering parameters, as a dataframe
}
