% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mortgage_state_allmonths_vectorised}
\alias{mortgage_state_allmonths_vectorised}
\title{mortgage_state_allmonths_vectorised}
\usage{
mortgage_state_allmonths_vectorised(
  rate,
  capital,
  term_months,
  month_num,
  month_overpay_vect = rep(0, month_num),
  reduce_term = TRUE
)
}
\arguments{
\item{rate}{double, the mortgage current rate (absolute value e.g. 1.5\% means rate = 0.015)}

\item{capital}{double, the mortgage remaining capital}

\item{term_months}{integer, number of months remaining to pay the mortgage back without overpayments}

\item{month_num}{integer, duration of payment period considered, in months}

\item{month_overpay_vect}{vector double, amount overpaid each month, defaults to rep(0,month_num)}

\item{reduce_term}{boolean, defaults to TRUE do overpayments reduce terms}
}
\value{
dataframe with each observation being one month and
variable including month, monthly_payment, overpayment, monthly_payment_tot, interest_paid,
interest_paid_tot and capital
}
\description{
shows state of mortgage as a dataframe with each observation being one month and
variable including month, monthly_payment, overpayment, monthly_payment_tot, interest_paid,
interest_paid_tot and capital
}
