% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageHelpers-package.R
\docType{package}
\name{MortgageHelpers-package}
\alias{MortgageHelpers}
\alias{MortgageHelpers-package}
\title{MortgageHelpers: Provide mortgage calculation helpers}
\description{
functions to calculate aspects of a mortgage including monthly payments, total terms and interest.
}
\keyword{internal}
