% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{price_fixedterm}
\alias{price_fixedterm}
\title{price_fixedterm}
\usage{
price_fixedterm(rate, capital, term_months, duration_years, fee = 0, rest = 12)
}
\arguments{
\item{rate}{double, the mortgage current rate (absolute value e.g. 1.5% means rate = 0.015)}

\item{capital}{double, the mortgage remaining capital}

\item{term_months}{integer, number of months remaining to pay the mortgage back}

\item{duration_years}{integer, duration of fixed term in years}

\item{fee}{double, the mortgage upfront fee}

\item{rest}{integer, defaults to 12 for monthly payments}
}
\value{

}
\description{
calculated the price of a fixed term of a mortgage
}
