% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{total_term_reduction_months}
\alias{total_term_reduction_months}
\title{total_term_reduction_months}
\usage{
total_term_reduction_months(
  rate,
  capital,
  term_months,
  month_overpay = 0,
  month_overpay_nums = 1:term_months
)
}
\arguments{
\item{rate}{double, the mortgage current rate (absolute value e.g. 1.5\% means rate = 0.015)}

\item{capital}{double, the mortgage remaining capital}

\item{term_months}{integer, number of months remaining to pay the mortgage back}

\item{month_overpay}{integer, amount overpaid each month, defaults to 0}

\item{month_overpay_nums}{vector, with index having non NA values will result in overpayment being applied for that month.
hence 1:month_num (default value) will mean overpayment for the whole payment period considered}
}
\value{
integer number of months removed from the mortgage duration
}
\description{
number of months removed from the mortgage duration compared to the case
where there is no overpayments
}
