% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{interest_paid_splitloan}
\alias{interest_paid_splitloan}
\title{interest_paid_splitloan}
\usage{
interest_paid_splitloan(
  rate,
  capital1,
  capital2,
  term_months,
  month_num,
  month_overpay = 0,
  perc_overpay1,
  perc_overpay2
)
}
\arguments{
\item{rate}{double, the mortgage current rate (absolute value e.g. 1.5% means rate = 0.015)}

\item{term_months}{integer, number of months remaining to pay the mortgage back}

\item{month_num}{integer, duration of payment period, in months, payments equal those required to pay off mortgage without overpayments}

\item{month_overpay}{integer, amount overpaid each month, defaults to 0}

\item{perc_overpay1}{percentage of overpay dedicated to first loan}

\item{perc_overpay2}{percentage of overpay dedicated to second loan}

\item{capital}{double, the mortgage remaining capital2}
}
\value{

}
\description{
interest paid if the loan is split in two with same rate.
}
